procfs: re-add limited caching for subset=pid
Commit 53d6de5e516a ("procfs: remove GLOBAL_PROCFS_HANDLE") removed the
caching of ProcfsHandles entirely. While it is true that fsopen(2) and
the rest of the mount API is relatively light-weight, some users will
make very heavy use of these APIs for every VFS operation and so having
a cache is really quite important (some tests with runc showed ~30% more
syscalls and a ~10% performance hit).

However, in contrast to GLOBAL_PROCFS_HANDLE, we will instead have
ProcfsHandle::new transparently cache the underlying OwnedFd if it is
considered "safe enough" (subset=pid and a detached mount object that
cannot trivially escape to the host with ".."). Users will not see any
visible difference (Drop will internally be a no-op for the cached
versions).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>