procfs: add ProcfsHandleRef
Now that we use MaybeOwnedFd in ProcfsHandle, we can adjust things so
that ProcfsHandleRef has a custom 'fd lifetime (tied to a BorrowedFd)
and ProcfsHandle has the 'static lifetime we use for the globally cached
handle.

This is mainly necessary for the C API to be able to take BorrowedFds as
an argument (necessary to allow C callers to amortise the cost of
getting an unmasked procfs).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>