procfs: add ProcfsHandleBuilder API
Rather than adding a new ProcfsHandle::new_* variant each time we want
to add a configuration option, we really should just bite the bullet and
go with a builder.

At the moment you can only really configure subset=pid but there are
some TODOs for future items. The plan is to wrap this in the C API so we
can add constructors for ProcfsHandle in the bindings.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>