capi: use symbol versions
I finally managed to figure out how to get this working. With this
slightly baroque setup, it is possible to have symbol versions that
work properly.

When we need to create backward-compatible functions, as long as the
version-script has the dependencies defined correctly you can just do

  #[no_mangle]
  pub extern "C" __pathrs_foobar_v1(...) -> ... { ... }
  utils::symver! {
      fn __pathrs_foobar_v1 <- (pathrs_foobar, version = "LIBPATHRS_0.1")
  }

and you will get a pathrs_foobar@LIBPATHRS_0.1 symbol. Unfortunately,
the __pathrs_foobar_v1 symbol will still be public, but at least by
prefixing it with "__" (and adding an exclude rule to cbindgen for it)
this should not be a huge issue.

I have also added aliases for the historical symbol names from libpathrs
0.1 as default symbols (...@@LIBPATHRS_0.1) both as a proof-of-concept
of how to avoid breakages and to avoid theoretical breakages from old
versions. They need to be default symbols because binaries built against
libpathrs 0.1 use the plain symbol names and so we need to mark the
symbols as being the default so the link loader can find them.

Ref: https://internals.rust-lang.org/t/support-symbol-versioning-with-export-name/23626
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>