go: move to 'cyphar.com/go-pathrs' vanity import
At the moment we have basically no Go users, so it's still safe for us
to make the switch now.

While I have used GitHub for more than a decade now, there is no
guarantee I will continue to use it forever and the way that Go (by
default) forces you to embed the hosting provider into your import
information in a way that cannot be migrated is quite frustrating.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>