cargo: remove hard max version requirements for dependencies
I added these so "cargo update" wouldn't update dependencies that would
break our MSRV, but making it a hard requirement was quite problematic.

In short, it would make it harder to use pathrs in other projects as you
would need to downgrade other dependencies to make pathrs happy (when
pathrs works fine with the newer versions, it's just an MSRV issue).
Ultimately, our CI will catch this, so remove the top-level
requirements.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>