contrib: add fake-enosys helper
The idea is quite simple -- generate a very minimal seccomp-bpf profile
which returns ENOSYS if the syscall matches the syscalls specified and
then execve(2) the provided command.

Luckily the (apparently unmaintained) bpfvm crate provides enough rope
to hang ourselves without needing to resort to libseccomp. We also don't
really need to deal with multiarch because we are only ever dealing with
the native architecture we run the tests on, so we can just use the raw
syscall numbers for our local architecture.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>