openat2: bump retry limit to 128
Based on some testing, the 32 limit can still fail in 3% of cases which
is a bit too high for some users. 128 only fails in ~0.1% of cases in
the worst-case "attacker is banging on the door" scenario, and higher
level runtimes that need stronger resiliency assurances can do extra
retry loops themselves by looking at ErrorKind.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>