procfs: use resolver for open_follow parentdir
We had assumed that the parent directory of ProcfsHandle::open_follow
was not a symlink (and so just did a standard ProcfsHandle::open
internally), which caused errors when running into that case. The
solution in that case is to just use the resolver to allow for regular
symlinks to be resolved when looking up the parent directory.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>