procfs: try to resolve all subpaths with resolver first
Our previous logic (switch to ProcfsHandle::open if readlink fails) had
several flaws which this patch addresses:

 1. For regular symlinks, we were trivially exploitable by overmount
    attacks. If a user tried to open /proc/mounts then a bind-mount over
    /proc/$pid/mounts or /proc/$pid would not be detected *at all*
    because we assumed all symlinks were magic-links and thus did not do
    any checking (not even fstype!).

 2. Somewhat similarly, if a parent component was a non-existent path or
    was a magic-link we would try ProcfsHandle::open again instead of
    just returning an error. This isn't a huge deal but it is needless
    duplication (given that readlink also does ProcfsHandle::open
    internally it was actually triplication of work).

Instead, we should try to do the resolution normally first (but without
forcefully setting O_NOFOLLOW). If this succeeds then the target path
was not a magic-link and did not contain any magic-link components (but
it may have contained symlink components, which we can resolve
normally). If this fails then we can fallback to the rest of
ProcfsHandle::open_follow. However, we only do the fallback for ELOOP
(which is the error you get from RESOLVE_NO_MAGICLINKS) -- all other
errors (including crossing mounts) are just returned.

This mostly goes without saying, but it bears repeating that this is not
safe against racing mounts -- an attacker could over-mount a magic-link
on top of a regular symlink and trigger ELOOP to then cause the fallback
to incorrectly open the symlink without validation. The window for this
might be tight (because we only do the no-validation symlink open for
ELOOP, so the bind-mount would need to be done after the mount ID is
checked but before readlink(2) is called).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>