procfs: opath resolver: make magic-link detection more comprehensive
Previously we would try to walk into certain magic-links (such as
"pipe:[...]", "<ns>:[...]", "anon_inode:[...]")  and just return ENOENT
which didn't match openat2(2)'s behaviour.

We cannot be certain that a particular symlink is a magic-link but we
can safely assume it is one if it looks like "foo:[bar]" as that is a
common pattern for magic-links to objects not in the regular filesystem.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>