go-pathrs: remove explicit pkg.go.dev links
Documentation links for Go identifiers are created implicitly,
i.e. there is no need to have an explicit link target.

In case it is not a standard library and there is no corresponding
import statement in the source file, the documentation link can be
created by specifying the full import path (as done here for Mknod).

Reference: https://go.dev/doc/comment#doclinks

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>