resolver: tests: skip /proc/net/unix for 'cargo test'
When running under 'cargo test', the procfs resolver tests fail when
dealing with /proc/net/unix because the thread-group leader can die
between us opening it and looking at d_path(). This results in these
tests failing because the path we see is actually end up seeing from
procfs is "/proc/$pid/net/unix (deleted)" (sometimes -- it's a race
condition, unfortunately).

For this problem we would normally switch to /proc/thread-self, but this
test is explicitly testing non-magic-link paths and so we need to just
skip this test for 'cargo test'.

Note that #[cfg(nextest)] is not an upstream thing and is something we
set in our test scripts, though I have opened a discussion[1] upstream
to propose it.

[1]: https://github.com/nextest-rs/nextest/discussions/2789

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>