openat2: more flexible fallback logic
Programs can configure seccomp-bpf filters during their execution,
potentially causing openat2(2) to start failing after libpathrs has
detected that it works.

The solution is two-fold:

(1) Only cache openat2(2) *failures* for the purposes of detecting if
    openat2(2) is supported.

    If a trivial openat2(2) fails then either the kernel does not
    support it or a seccomp-bpf filter was activated. In either case,
    the program cannot revert to a state where openat2(2) is supported
    again, so it's safe to cache this.

    On the other hand, openat2(2) working in the past does not guarantee
    it will continue to work, so we cannot cache this result.

(2) Because the resolver type is picked at the creation time of the
    top-level object being used for resolutions (ProcfsHandle and Root),
    the above solution is not enough, as objects created before the
    seccomp-bpf profile are created could have the wrong resolver.

    The solution here is a little more annoying -- the only option is to
    do a fallback if openat2(2) is not supported at the time we try to
    use the resolver. There are a few options here (with different
    downsides), but the goal should be to be as efficient as possible
    for the openat2(2)-is-supported case.

    To wit, for the openat2(2) resolvers we always try the actual
    openat2(2) operation first. If this works, then we just return the
    value and everything worked out. If it fails, then we check if any
    openat2(2) call will fail (by trying a basic open-"." call) -- if
    this is the case then we do the fallback, otherwise we return the
    error from openat2(2) directly. This secondary check is necessary to
    make sure that we don't fall victim to a downgrade attack if an
    attacker finds a bug in the O_PATH resolver that triggers an error
    in the openat2(2) resolver.

All of the above was already done in pathrs-lite (though only for the
Root resolver -- the procfs resolver doesn't do this, but in this patch
to libpathrs we do it for all resolvers).

However, one unfortunate aspect of libpathrs is that the resolvers are
all enums with an "impl Default" that selects the most appropriate
resolver automatically. As per (1), this would naively require every
Default call to do a dummy openat2(2) call, which is needlessly
expensive (especially with RootRef through C FFI -- every pathrs_*
function call will do an extra openat2(2) call first).

The solution is for this case to only check if there is a cached failure
without attempting openat2(2) if there isn't. If there is a cached
failure we can just select O_PATH by default, otherwise we select
openat2(2) so that if it is supported then we have no extra overhead.
Otherwise, the check for support is deferred to the error path, as
intended by (2).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>